volviendo

+++ src/main/java/com/mandarinaSolutions/impresiones3d/repository/Repository.java
@@ -4,8 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mandarinaSolutions.impresiones3d.dominio.Articulo;
 
-public interface Repository extends JpaRepository<Articulo, Long> {
-	
-	
+public interface Repository extends JpaRepository<Articulo, Integer> {
 
 }
